refactor: fragmentado o código em componentes
